+++ pkg/proxy/winkernel/hcntesting/fakehcn.go
@@ -2,6 +2,7 @@ package hcntesting
 
 import (
 	"encoding/json"
+	"errors"
 
 	"github.com/Microsoft/hcsshim/hcn"
 	"k8s.io/klog/v2"
@@ -99,7 +100,12 @@ func (HCN *FakeHCN) GetEndpointByID(endpointId string) (*hcn.HostComputeEndpoint
 			endpoint.Flags = ep.Flags
 			endpoint.Health = ep.Health
 			endpoint.SchemaVersion = ep.SchemaVersion
+			break
 		}
+	}
+	if endpoint.Id == "" {
+		err := errors.New("No endpoint matches this ID")
+		return nil, err
 	}
 	return endpoint, nil
 }
@@ -129,7 +135,12 @@ func (HCN *FakeHCN) GetEndpointByName(endpointName string) (*hcn.HostComputeEndp
 			endpoint.MacAddress = ep.MacAddress
 			endpoint.Flags = ep.Flags
 			endpoint.SchemaVersion = ep.SchemaVersion
+			break
 		}
+	}
+	if endpoint.Name == "" {
+		err := errors.New("No endpoint matches this NAME")
+		return nil, err
 	}
 	return endpoint, nil
 }
@@ -153,7 +164,12 @@ func (HCN *FakeHCN) GetLoadBalancerByID(loadBalancerId string) (*hcn.HostCompute
 			loadbalancer.SchemaVersion = lb.SchemaVersion
 			loadbalancer.PortMappings = lb.PortMappings
 			loadbalancer.FrontendVIPs = lb.FrontendVIPs
+			break
 		}
+	}
+	if loadbalancer.Id == "" {
+		err := errors.New("No loadBalancer matches this ID")
+		return nil, err
 	}
 	return loadbalancer, nil
 }
@@ -228,7 +244,7 @@ func (HCN *FakeHCN) DeleteLoadBalancer(loadbalancer *hcn.HostComputeLoadBalancer
 
 	i--
 
-	if len(HCN.Loadbalancers) != 0 {
+	if len(HCN.Loadbalancers) != 0 { //here we actually delete the load balancer from fakehcn memory
 		copy(HCN.Loadbalancers[i:], HCN.Loadbalancers[i+1:])
 		HCN.Loadbalancers[len(HCN.Loadbalancers)-1] = nil
 		HCN.Loadbalancers = HCN.Loadbalancers[:len(HCN.Loadbalancers)-1]
@@ -249,7 +265,7 @@ func (HCN *FakeHCN) DeleteEndpoint(endpoint *hcn.HostComputeEndpoint) error {
 
 	i--
 
-	if len(HCN.Endpoints) != 0 {
+	if len(HCN.Endpoints) != 0 { //here we actually delete the endpoint from fakehcn memory
 		copy(HCN.Endpoints[i:], HCN.Endpoints[i+1:])
 		HCN.Endpoints[len(HCN.Endpoints)-1] = nil
 		HCN.Endpoints = HCN.Endpoints[:len(HCN.Endpoints)-1]
